Switched from identity to value comparisons (is verus ==).
Identify comparisions of strings may or may not match in python, since the
second string can be a different instance with the same value. The frequency
at which this happens depends on the interpreter.

This was causing library failures for me, so I fixed it. I made similar
changes to boolean comparisons, since I wasn't sure if all boolean values share
identity or not.